Add a option to enable PK chunking w/the bulk API
Add a `pk_chunking` boolean config option to use PK chunking from the
start, instead of failing over to it when a job fails w/a query timeout.

This is useful when syncing large tables. The query timeout behavior
has been preserved, but does not seem to work in all scenarios.